Simplified the social preview rendering
