+++ webpack/paths.js
@@ -33,7 +33,6 @@ const entry = {
 	"wp-seo-filter-explanation": "./wp-seo-filter-explanation.js",
 	"wp-seo-quick-edit-handler": "./wp-seo-quick-edit-handler.js",
 	"wp-seo-network-admin": "./wp-seo-network-admin.js",
-	"wp-seo-edit-social": "./edit-social.js",
 };
 
 /**
